Abstract router to index.js
