+++ views/public/index.js
@@ -47,7 +47,7 @@ function submitPlan(userId, planName, course1, course2, course3, course4,
   course5, course6, course7, course8, course9, course10, course11, course12) {
 
   const postRequest = new XMLHttpRequest();
-  const postURL = "/appliedplanportal/plan";
+  const postURL = "/plan";
   postRequest.open("POST", postURL);
 
   const postObj = {
